Font Change, Respawn Menu, and Different Scene For Character Selection
The font was changed to make it look much more clean and nice. A respawn menu was added so you no longer just go to the main menu but instead are put to a special screen. The character selection screen was put to a different scene to make the inventory work.